refactor: moved spawn trigger to the chunk load listener
